Fixed issue with integer values being misinterpreted by glsl by using glVertexArribIPointer for specific data types. Won't soon be forgetting that one.

+++ Renaissance/assets/shaders/Texture.glsl
@@ -4,13 +4,13 @@
 layout(location=0) in vec3 a_Pos;
 layout(location=1) in vec4 a_Tint;
 layout(location=2) in vec2 a_TexCoord;
-layout(location=3) in float a_TexIndex;
+layout(location=3) in uint a_TexIndex;
 
 uniform mat4 u_ViewProjection;
 
 out vec4 v_Tint;
 out vec2 v_TexCoord;
-flat out float v_TexIndex;
+flat out uint v_TexIndex;
 
 void main()
 {
@@ -25,7 +25,7 @@ void main()
 
 in vec4 v_Tint;
 in vec2 v_TexCoord;
-flat in float v_TexIndex;
+flat in uint v_TexIndex;
 
 layout (binding=0) uniform sampler2D u_Textures[16];
 
@@ -33,5 +33,5 @@ out vec4 f_Col;
 
 void main()
 {
-	f_Col = texture(u_Textures[int(v_TexIndex)], v_TexCoord) * v_Tint;
+	f_Col = texture(u_Textures[v_TexIndex], v_TexCoord) * v_Tint;
 }
